changed photo's sequence to long

+++ src/main/java/com/beeswork/balanceaccountservice/dto/account/PhotoDTO.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Setter
 public class PhotoDTO {
     private String key;
-    private int sequence;
+    private Long sequence;
 }
